requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
redis==5.0.1
PyPDF2==3.0.1
google-generativeai==0.3.2
requests==2.31.0
pydantic==2.5.0